Move PropertyGrid extensions to custom project
